Deprecated geo::PlaneGeo::SignalType() .
This is toward the resolution of issue #14365 .
Also geo::PlaneGeo::SetSignalType() is to be considered deprecated, but it's not marked so because it is still necessary to maintain the legacy, deprecated functionality.
Instead, it's being made private (it's not supposed to be used except than for post-processing during geometry initialisation).