Fix on one of geo::GeometryCore::WireIDsIntersect() methods.
This should solve issue #17123.

A new implementation is coded, after the suggestion in the ticket.
The unit test has been extended to check the reverse intersection
(that is, both (w1,w2) and (w2,w1)). Also a test of consistency with the
legacy implementation has been added when drift direction is +/-x.
There is still a risk that the unit test is too self-righteous and fails to
detect errors.